Update the ability to go to the transfer page when you detect a shake in the background
